refactor: refactor TimesNet and fix some typos in its doc;
